deleted excess folders and strange duplicates
