+++ Frontend/src/Hooks/useLogin.jsx
@@ -6,8 +6,8 @@ export const useLogin = () =>{
     const [isLoading, setIsLoading] = useState(false)
     const [error, setError] = useState(null) 
     // const {dispatch} = useAuthContext()
-
-
+ 
+ 
     const login = async (email, password) => {
         if(!email || !password)
         {
@@ -17,7 +17,7 @@ export const useLogin = () =>{
             setIsLoading(true)
             setError(null)
             console.log("Is here")
-            const response = await fetch('/user/login',{
+            const response = await fetch('http://localhost:8000/user/login',{
                 method:'POST',
                 headers: {'Content-Type': 'application/json'},
                 body:JSON.stringify({email, password })
