+++ Frontend/src/Pages/LoginPage.jsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { Link } from 'react-router-dom';
-// import {useLogin} from '../Hooks/useLogin'
+import {useLogin} from '../Hooks/useLogin'
 
 import '../Styles/LoginPage.css'
 const LoginPage = () => {
@@ -8,13 +8,13 @@ const LoginPage = () => {
     const [password, setPassword] = useState('')
     const [emptyFields, setEmptyFields] = useState([])
 
-    // const {error, isLoading, login} = useLogin()
+    const {error, isLoading, login, setError} = useLogin()
 
-    const handleSubmit = (e) =>{
-        e.preventDefault()
+    const handleSubmit = (e) =>{ 
+        e.preventDefault() 
         setEmptyFields([])
-
-        // login(email, password)
+        setError(null)
+        login(email, password)
     }
     return (  
         <div className='user-login-screen'>
@@ -33,10 +33,9 @@ const LoginPage = () => {
 
                     <h3>Forgot your password? Click <a href="/">Here</a></h3> <br />
 
-                    {/* {error && <div className="error">{error}</div> } */}
+                    {error && <div className="error">{error}</div> }
  
-                    <button >Login</button>
-                    {/* <button disabled = {isLoading} >Login</button> */}
+                    <button disabled={isLoading}>Login</button>
                     
                 </form>
             </div>
